careeropen-backend/requirements-clean.txt
# Core Django
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0,<25.0

# Database
psycopg2-binary>=2.9,<3.0
redis>=5.0,<6.0
django-redis>=5.4.0,<6.0.0

# API Documentation
drf-spectacular==0.28.0

# Configuration
python-decouple>=3.8,<4.0

# Security
django-cors-headers>=4.3,<5.0

# Storage
boto3>=1.34,<2.0
django-storages>=1.14,<2.0
whitenoise>=6.6,<7.0

# Production
gunicorn>=22.0,<23.0

# Testing
pytest>=8.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
factory-boy>=3.3.0
freezegun>=1.2.0

# Development
ipython>=8.0.0
django-debug-toolbar>=4.0.0
